server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
configparser
psycopg2
pyjwt
flask_jwt_extended